Rename all namespaces
