+++ src/Heuristic.Linq/ComparerExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace LinqToAStar
+namespace Heuristic.Linq
 {
     static class ComparerExtensions
     {
