+++ src/Heuristic.Linq/DistanceHelper.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace LinqToAStar
+namespace Heuristic.Linq
 {
     /// <summary>
     /// Provides a set of static methods that calculate the distance between two positions.
