+++ src/Heuristic.Linq/HeuristicSearchSelect.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace LinqToAStar
+namespace Heuristic.Linq
 {
     internal class HeuristicSearchSelect<TSource, TFactor, TStep> : HeuristicSearchBase<TFactor, TStep>
     {
